Add interfaces inc. Handler + impl. Pipe,HandlerFunc to stream package
The stream.Handler interface is intended to model streaming behavior, on either
the client or server. The stream.PipeReader and stream.PipeWriter interfaces
match the signatures of io.PipeReader and io.PipeWriter. The stream.Pipe
implementation supports attaching an io.Closer to either / both of those,
and supports joining one side of a bi-directional stream into a single
io.ReadWriteCloser, and also implements both stream.PipeReader and
stream.PipeWriter.