Add numerous implementations cumulating in ionet.Wrap and stream.HalfCloser
The purpose of ionet.Wrap is to make it easier to implement net.Conn using
arbitrary streams. The purpose of stream.HalfCloser is to support patterns
that "send EOF" (like unix pipes), specifically for use in implementations
where that operation uses an underlying communication mechanism / data stream
as the write, and is therefore difficult to use safely in way that is
compatible with the common usage of io.ReadWriteCloser.

Other notable inclusions are ionet.Pipe, an extension of net.Pipe that exposes
the underlying stream.PipeReader and stream.PipeWriter pairs, and
stream.SyncPipe, which (while limited) is a fairly reliable and relatively safe
mechanism that facilitates "calling close after writing all the data", even
when there's pipe/copy operations between the write and wherever the close
terminates.