Refactor stream.HalfCloser eliminate certain false positive full-close cases
The aim of this change is to ensure that writes which are initiated after close
(of the half closer), where it's not forcibly closing the full pipe, will not
have a chance of triggering a full close. The race was occuring very
occasionally, between acquiring the write mutex in the close impl., and the
deferred `closePipe(false)` call, the final step in the body of the close func.