First pass graceful closer support for remote buffered e.g. proxied streams
In order to support stream proxying, it is necessary to send EOF, then wait
for all data to processed, including all data to be read, before freeing the
associated resources. This change adds a variety of tooling to better support
this, including ionet.WrapPipeGraceful, stream.GracefulCloser, and the
CloseGuarder and GracefulCloser options for stream.HalfCloser (and the
ionet.WrapPipe* functions).