Fix stream.HalfCloser option GracefulCloser deadlock don't use internally
Since the graceful closer may involve waiting _on the halfcloser itself_ (e.g.
ionet.WrapPipeGraceful), the halfcloser "full close" case could result in a
deadlock, as this case may occur WITHIN the sync.Once. This change simply
prevents that from happening. The graceful closer was only relevant for
external callers, in any case, as the graceful case for stream.HalfCloser does
not involve calling that closer, in the first place.