Fix tun/grpc unsafe half close behavior
Adds a wrapper implementation that synchronises the client message sends, with
support for cancelation of lower-priority messages, and avoidance of multiple
cancels or half closes.

This also changes the type of the stream id to uint64, and tightens up other
aspects of both the existing implementation and the associated schema.

The unit testing is not adequate, so there may be significant bugs.